requirements.txt
Chroma                  == 0.2.0
chromadb                == 0.6.3
cmake                   == 3.31.4
faiss-cpu               == 1.10.0
git-filter-repo         == 2.47.0
grpcio-status           == 1.70.0
httptools               == 0.6.4
jurigged                == 0.6.0
langchain-cli           == 0.0.35
langchain-community     == 0.3.16
langchain-faiss         == 0.1.1
langchain-google-genai  == 2.0.9
langchain-groq          == 0.2.4
langchain-huggingface   == 0.1.2
langchain-pinecone      == 0.2.3
openpyxl                == 3.1.5
pandas                  == 2.2.3
pdfminer                == 20191125
pinecone-client         == 3.1.0
pip                     == 25.0.1
pymupdf4llm             == 0.0.17
python-multipart        == 0.0.20
redis                   == 5.2.1
setuptools              == 75.8.0
shellingham             == 1.5.4
sse-starlette           == 1.8.2
unstructured            == 0.16.20
watchfiles              == 1.0.4
websockets              == 14.2
wheel                   == 0.45.1
youtube-transcript-api  == 0.6.3